requirements.txt
# YOLO / Object Detection
ultralytics>=8.0.0

# PyTorch (CPU/GPU)
torch>=2.0.0
torchvision>=0.15.0

# Image / Video processing
opencv-python>=4.7.0
matplotlib>=3.7.0
numpy>=1.25.0
pandas>=2.0.0
tqdm>=4.65.0

# Text/keyboard processing
pyautogui

# Optional: for dataset handling
PyYAML>=6.0
